requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
Babel==2.12.1
beautifulsoup4==4.12.0
billiard==3.6.4.0
cachetools==4.2.4
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==3.0.0
Django==3.2.11
django-allauth==0.52.0
django-braces==1.15.0
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-cleanup==7.0.0
django-countries==7.5.1
django-environ==0.10.0
django-extensions==3.2.1
django-instagram-profile==1.0.3
django-json-field==0.5.7
django-json-widget==1.1.1
django-oauth-toolkit==2.2.0
django-phonenumber-field==2.0.1
django-rename-app==0.1.5
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==4.2.0
django-rest-passwordreset==1.3.0
django-sslserver==0.22
django-templated-mail==1.1.1
django-timezone-field==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.5
drf-yasg2==1.19.4
EasyProcess==1.1
entrypoint2==1.1
facebook-sdk==3.1.0
flower==1.2.0
future==0.18.3
google==3.0.0
google-api-core==1.22.2
google-api-python-client==1.12.1
google-auth==1.21.2
google-auth-httplib2==0.0.4
googleapis-common-protos==1.52.0
httplib2==0.22.0
humanize==4.6.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
kombu==5.2.4
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.0
pandas==1.5.3
patool==1.12
phonenumbers==8.13.6
Pillow==9.4.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==4.22.1
psycopg2==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.1.0
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-twitter==3.5
python3-openid==3.2.0
pytz==2022.7.1
pyunpack==0.3
PyYAML==6.0
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scipy==1.10.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
soupsieve==2.4
sqlparse==0.4.3
tornado==6.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
